# swagger.yaml
openapi: 3.0.3
info:
  title: MeetFlow API
  version: 1.0.0
  description: |
    ## MeetFlow Backend REST API

    This API powers the MeetFlow SaaS platform. It is divided into four
    top-level sections:

    | Section          | Base Path          | Description                                    |
    |------------------|--------------------|------------------------------------------------|
    | **Common**       | `/api/v1/auth` & `/api/v1/account`          | Authentication & account management            |
    | **Master**       | `/api/v1/master`   | MeetFlow internal admin panel (master context) |
    | **Organization** | `/api/v1/organization` | Organisation-level management (coming soon) |
    | **Workspace**    | `/api/v1/workspace`    | Workspace & channel management (coming soon)|

    ### Authentication
    All **Master**, **Organization**, and **Workspace** routes require a valid
    JWT `accessToken` in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```
    Use the **Login** endpoint to obtain tokens.

    ### Standard Response Envelope
    Every response (success or error) is wrapped in:
    ```json
    {
      "status": {
        "response_code": 200,
        "response_message": "Success!"
      },
      "message": "Human-readable message",
      "data": { ... }
    }
    ```
    Error responses replace `data` with an `error` object containing `details`.

  contact:
    name: MeetFlow Engineering
  license:
    name: ISC
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token returned by POST /auth/login.
        Valid for 1 day (or 30 days when rememberMe is true).
...
  schemas:
...
    StatusBlock:
      type: object
      properties:
        response_code:
          type: integer
          example: 200
        response_message:
          type: string
          example: "Success!"
    SuccessEnvelope:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/StatusBlock"
        message:
          type: string
          example: "Operation successful"
        data:
          description: Payload — varies per endpoint
          nullable: true
    ErrorEnvelope:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/StatusBlock"
        message:
          type: string
          example: "Something went wrong!"
        error:
          type: object
          properties:
            details:
              type: string
              example: "Invalid credentials."
    DeletedResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              nullable: true
              example: null

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    PlanFeature:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        plan_id:
          type: integer
          format: int64
          example: 2
        feature_id:
          type: integer
          format: int64
          example: 3
        is_active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
    AssignFeatureRequest:
      type: object
      required:
        - feature_id
      properties:
        feature_id:
          type: integer
          format: int64
          description: ID of the feature to assign
          example: 3
        is_active:
          type: boolean
          default: true
    ToggleFeatureRequest:
      type: object
      required:
        - is_active
      properties:
        is_active:
          type: boolean
          example: false
          description: Set true to enable, false to disable the feature on this plan

    # ── Plan Limit ────────────────────────────────────────────────────────────
...
    BulkAssignPlanFeatureItem:
      type: object
      required:
        - feature_id
      properties:
        feature_id:
          type: integer
          format: int64
          example: 3
        is_active:
          type: boolean
          default: true
    BulkAssignPlanFeatureRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/BulkAssignPlanFeatureItem"
    BulkRemovePlanFeatureRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: integer
            format: int64
          example: [1, 2, 3]

    # ── Bulk Plan Limits ──────────────────────────────────────────────────────
...
  responses:
    UnauthorizedError:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            status:
              response_code: 401
              response_message: "Something went wrong!"
            message: "Access token is required."
    ForbiddenError:
      description: Authenticated but insufficient permissions (wrong context)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            status:
              response_code: 403
              response_message: "Something went wrong!"
            message: "This route is only accessible to master users."
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequestError:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
paths:
...
  /master/plans/get-plan-features/{planId}:
    get:
      tags: [Master / Plan Features]
      summary: Get features assigned to a plan
      operationId: getFeaturesByPlan
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
      responses:
        "200":
          description: Plan features retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/PlanFeature"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/plans/assign-plan-feature/{planId}:
    post:
      tags: [Master / Plan Features]
      summary: Assign a feature to a plan
      operationId: assignFeature
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignFeatureRequest"
            example:
              feature_id: 3
              is_active: true
      responses:
        "201":
          description: Feature assigned successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PlanFeature"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/plans/toggle-plan-feature/{planId}/{featureId}:
    patch:
      tags: [Master / Plan Features]
      summary: Toggle a feature on a plan (enable / disable)
      operationId: toggleFeature
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
        - name: featureId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleFeatureRequest"
            example:
              is_active: false
      responses:
        "200":
          description: Feature toggled successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PlanFeature"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/plans/remove-plan-feature/{planId}/{featureId}:
    delete:
      tags: [Master / Plan Features]
      summary: Remove a feature from a plan
      operationId: removeFeature
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
        - name: featureId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 3
      responses:
        "200":
          description: Feature removed from plan successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ─── Bulk Plan Feature Operations ──────────────────────────────────────────
  /master/plans/bulk-assign-features/{planId}:
    post:
      tags: [Master / Plan Features]
      summary: Bulk assign features to a plan
      description: |
        Assign multiple features to a plan in a single request.
        **Limits:** Maximum 100 items. Each feature_id must be unique in the request.
      operationId: bulkAssignFeatures
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          description: ID of the plan
          schema:
            type: integer
            format: int64
            example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAssignPlanFeatureRequest"
            example:
              items:
                - feature_id: 1
                  is_active: true
                - feature_id: 2
                  is_active: true
                - feature_id: 3
                  is_active: false
      responses:
        "200":
          description: Bulk assign completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          successful:
                            type: array
                            items:
                              $ref: "#/components/schemas/PlanFeature"
                          failed:
                            type: array
                          summary:
                            type: object
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/plans/bulk-remove-features/{planId}:
    delete:
      tags: [Master / Plan Features]
      summary: Bulk remove features from a plan
      description: |
        Remove multiple features from a plan by feature IDs.
        **Limits:** Maximum 100 feature IDs per request.
      operationId: bulkRemoveFeatures
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          description: ID of the plan
          schema:
            type: integer
            format: int64
            example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkRemovePlanFeatureRequest"
            example:
              ids: [1, 2, 3]
      responses:
        "200":
          description: Bulk remove completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ── Plan Limits (sub-resource) ────────────────────────────────────────────
...